+++ Xplat/src/main/java/vazkii/botania/common/block/LuminizerBlock.java
@@ -28,7 +28,6 @@ import net.minecraft.world.level.block.entity.BlockEntityType;
 import net.minecraft.world.level.block.state.BlockState;
 import net.minecraft.world.level.block.state.StateDefinition;
 import net.minecraft.world.level.block.state.properties.BlockStateProperties;
-import net.minecraft.world.level.gameevent.GameEvent;
 import net.minecraft.world.phys.BlockHitResult;
 import net.minecraft.world.phys.shapes.CollisionContext;
 import net.minecraft.world.phys.shapes.VoxelShape;
@@ -36,11 +35,10 @@ import net.minecraft.world.phys.shapes.VoxelShape;
 import org.jetbrains.annotations.NotNull;
 import org.jetbrains.annotations.Nullable;
 
-import vazkii.botania.api.internal.VanillaPacketDispatcher;
 import vazkii.botania.api.state.enums.LuminizerVariant;
 import vazkii.botania.common.block.block_entity.BotaniaBlockEntities;
 import vazkii.botania.common.block.block_entity.LuminizerBlockEntity;
-import vazkii.botania.common.item.BotaniaItems;
+import vazkii.botania.common.item.PhantomInkItem;
 
 public class LuminizerBlock extends BotaniaWaterloggedBlock implements EntityBlock {
 
@@ -70,15 +68,7 @@ public class LuminizerBlock extends BotaniaWaterloggedBlock implements EntityBlo
 		ItemStack stack = player.getItemInHand(hand);
 		BlockEntity te = world.getBlockEntity(pos);
 		if (te instanceof LuminizerBlockEntity relay) {
-			if (stack.is(BotaniaItems.phantomInk) && !relay.isNoParticle()) {
-				if (!world.isClientSide) {
-					stack.shrink(1);
-					relay.setNoParticle();
-					world.gameEvent(null, GameEvent.BLOCK_CHANGE, pos);
-					VanillaPacketDispatcher.dispatchTEToNearbyPlayers(relay);
-				}
-				return InteractionResult.sidedSuccess(world.isClientSide());
-			} else if (!stack.is(Items.ENDER_PEARL)) {
+			if (!stack.is(Items.ENDER_PEARL) && !(stack.getItem() instanceof PhantomInkItem)) {
 				relay.mountEntity(player);
 				return InteractionResult.sidedSuccess(world.isClientSide());
 			}