+++ Xplat/src/main/java/vazkii/botania/common/item/BotaniaItems.java
@@ -313,7 +313,7 @@ public final class BotaniaItems {
 	public static final Item blackerLotus = make(prefix(LibItemNames.BLACKER_LOTUS), new BlackLotusItem(defaultBuilder().rarity(Rarity.EPIC)));
 	public static final Item worldSeed = make(prefix(LibItemNames.WORLD_SEED), new WorldSeedItem(defaultBuilder()));
 	public static final Item overgrowthSeed = make(prefix(LibItemNames.OVERGROWTH_SEED), new OvergrowthSeedItem(defaultBuilder().rarity(Rarity.RARE)));
-	public static final Item phantomInk = make(prefix(LibItemNames.PHANTOM_INK), new Item(defaultBuilder()));
+	public static final Item phantomInk = make(prefix(LibItemNames.PHANTOM_INK), new PhantomInkItem(defaultBuilder()));
 	public static final Item poolMinecart = make(prefix(LibItemNames.POOL_MINECART), new ManaPoolMinecartItem(unstackable()));
 	public static final Item keepIvy = make(prefix(LibItemNames.KEEP_IVY), new ResoluteIvyItem(defaultBuilder()));
 	public static final Item placeholder = make(prefix(LibItemNames.PLACEHOLDER), new SelfReturningItem(defaultBuilder()));